Updated Run Method to add NameSpace Array (#23)
* Update RequestSchema.php

Updated __construct to accept 2nd parameter that qualifies the first param as either as a File or JSON. The 2nd parameter defaults to "file" so existing functionality should not break.

* Update MessageTranslator.php

Updated 
    public function setTranslationTable($input,$inputtype='file'){
and
    public function setDefaultTable($input,$inputtype='file'){

To receive a 2nd parameter as that qualifies the first one to be either a File or an array. 

Default is file so this will not break existing functionality.

* Removed input type param, added addTranslator

Removed the 2nd param for input type and added addTranslationTable and addDefaultTable methods with exception handling

* Fixed issue, Added addSchema function

Fixed issue with __constructor and added addSchema function. Still testing this.

* Added Form Array Namespace

Added an additional parameter to the rules method to add array prefix for all field names. So the validadtion rules match the field names in the form.

* Updated Field name in the Messages

* Updated Readme to add Form Namespace

Text editor also formatted the text a little, so it shows changes but nothing else was changed.